Demo增加周K和月

+++ klinelib/src/main/java/com/guoziwei/klinelib/model/HisData.java
@@ -10,9 +10,9 @@ public class HisData {
     private double high;
     private double low;
     private double open;
-    private int vol;
+    private long vol;
     private long date;
-    private int amountVol;
+    private long amountVol;
     private double avePrice;
     private double total;
     private double maSum;
@@ -79,11 +79,11 @@ public class HisData {
         this.open = open;
     }
 
-    public int getVol() {
+    public long getVol() {
         return vol;
     }
 
-    public void setVol(int vol) {
+    public void setVol(long vol) {
         this.vol = vol;
     }
 
@@ -97,11 +97,11 @@ public class HisData {
     }
 
 
-    public int getAmountVol() {
+    public long getAmountVol() {
         return amountVol;
     }
 
-    public void setAmountVol(int amountVol) {
+    public void setAmountVol(long amountVol) {
         this.amountVol = amountVol;
     }
